requirements.txt
djangorestframework==3.15.1
django-cors-headers==4.3.1
pymysql==1.1.0
python-dotenv==1.0.1
mysqlclient==2.1.1